Don't show the current opensearch engine icon in the toolbar search. This allows us to not have to load the networkaccessmanager on startup which is a large cost.  We can always add this later and users will probably not complain if we add it, but they will if we remove it because we are unable to find a good enough solution.  Also by not showing the icon we can also delay working on adding a down arrow to the open search icons in the toolbar search.
In this same vain change OpenSearchEngine to not load the engine icons
until they are needed.  Before this they were all being fetched on
startup which would be an ever growing startup cost.  After the first
load they will be cached and will show up _very_ quickly.

Added a new QAction subclass that can automatically update its icon
when the engine gets it.  This way the engine image doesn't need
to be loaded before it is needed in a menu.